fix: split creation page under /create, fix posthog provider so refresh works
