+++ app/page.tsx
@@ -2,324 +2,15 @@
 
 import { ViewType } from '@/components/auth'
 import { AuthDialog } from '@/components/auth-dialog'
-import { Chat } from '@/components/chat'
-import { ChatInput } from '@/components/chat-input'
-import { ChatPicker } from '@/components/chat-picker'
-import { CreatorInput } from '@/components/creator-input'
-import { NavBar } from '@/components/navbar'
-import { Preview } from '@/components/preview'
 import { Header, Hero, Features, HowItWorks, CTASection, Footer } from '@/components/landing'
 import { useAuth } from '@/lib/auth'
-import {
-  Message,
-  sanitizeMessagesForStorage,
-  toAISDKMessages,
-  toMessageImage,
-  toMessageFile,
-} from '@/lib/messages'
-import { LLMModelConfig } from '@/lib/models'
-import modelsList from '@/lib/models.json'
-import { FragmentSchema, fragmentSchema as schema } from '@/lib/schema'
 import { supabase } from '@/lib/supabase'
-import templates, { getTemplateIdSuffix } from '@/lib/templates'
-import { ExecutionResult } from '@/lib/types'
-import { DeepPartial } from 'ai'
-import { experimental_useObject as useObject } from '@ai-sdk/react'
-import { usePostHog } from 'posthog-js/react'
-import { SetStateAction, useEffect, useState, useRef } from 'react'
-import { useLocalStorage } from 'usehooks-ts'
-import { ArrowLeft } from 'lucide-react'
+import { useState } from 'react'
 
 export default function Home() {
-  const [chatInput, setChatInput] = useLocalStorage('chat', '')
-  const [files, setFiles] = useState<File[]>([])
-  const [pdfFiles, setPdfFiles] = useState<File[]>([])
-  const [selectedTemplate, setSelectedTemplate] = useState<string>(
-    getTemplateIdSuffix('explorable-research-developer'),
-  )
-  const [languageModel, setLanguageModel] = useLocalStorage<LLMModelConfig>(
-    'languageModel',
-    {
-      model: 'google/gemini-3-pro-preview:online',
-    },
-  )
-
-  const posthog = usePostHog()
-
-  const [result, setResult] = useState<ExecutionResult>()
-  const [messages, setMessages] = useState<Message[]>([])
-  const messagesRef = useRef<Message[]>([])
-  const [fragment, setFragment] = useState<DeepPartial<FragmentSchema>>()
-  const [currentTab, setCurrentTab] = useState<'code' | 'fragment'>('code')
-  const [isPreviewLoading, setIsPreviewLoading] = useState(false)
   const [isAuthDialogOpen, setAuthDialog] = useState(false)
   const [authView, setAuthView] = useState<ViewType>('sign_in')
-  const [isRateLimited, setIsRateLimited] = useState(false)
-  const [errorMessage, setErrorMessage] = useState('')
-  const [showChatInterface, setShowChatInterface] = useState(false)
-  const { session, userTeam } = useAuth(setAuthDialog, setAuthView)
-  const [useMorphApply, setUseMorphApply] = useLocalStorage(
-    'useMorphApply',
-    process.env.NEXT_PUBLIC_USE_MORPH_APPLY === 'true',
-  )
-
-  const chatSectionRef = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    messagesRef.current = messages
-  }, [messages])
-
-  async function saveProject({
-    fragment,
-    result,
-  }: {
-    fragment: DeepPartial<FragmentSchema> | undefined
-    result: ExecutionResult | undefined
-  }) {
-    if (!session?.access_token || !fragment || !result) {
-      return
-    }
-
-    const sanitizedMessages = sanitizeMessagesForStorage(
-      messagesRef.current ?? [],
-    )
-    const updatedMessages = sanitizedMessages.map((message, index) => {
-      if (
-        index === sanitizedMessages.length - 1 &&
-        message.role === 'assistant'
-      ) {
-        return { ...message, object: fragment, result }
-      }
-      return message
-    })
-
-    try {
-      const response = await fetch('/api/projects', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${session.access_token}`,
-        },
-        body: JSON.stringify({
-          fragment,
-          result,
-          messages: updatedMessages,
-        }),
-      })
-
-      if (!response.ok) {
-        throw new Error('Failed to save project')
-      }
-    } catch {
-      console.error('Failed to save project')
-    }
-  }
-
-  const filteredModels = modelsList.models.filter((model) => {
-    if (process.env.NEXT_PUBLIC_HIDE_LOCAL_MODELS) {
-      return model.providerId !== 'ollama'
-    }
-    return true
-  })
-
-  const defaultModel = filteredModels.find(
-    (model) => model.id === 'google/gemini-3-pro-preview:online',
-  ) || filteredModels[0]
-
-  const currentModel = filteredModels.find(
-    (model) => model.id === languageModel.model,
-  ) || defaultModel
-
-  // Update localStorage if stored model no longer exists
-  useEffect(() => {
-    if (languageModel.model && !filteredModels.find((m) => m.id === languageModel.model)) {
-      setLanguageModel({ ...languageModel, model: defaultModel.id })
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [languageModel.model])
-  
-  const currentTemplate =
-    selectedTemplate === 'auto'
-      ? templates
-      : { [selectedTemplate]: templates[selectedTemplate] }
-  const lastMessage = messages[messages.length - 1]
-
-  // Determine which API to use based on morph toggle and existing fragment
-  const shouldUseMorph =
-    useMorphApply && fragment && fragment.code && fragment.file_path
-  const apiEndpoint = shouldUseMorph ? '/api/morph-chat' : '/api/chat'
-
-  const { object, submit, isLoading, stop, error } = useObject({
-    api: apiEndpoint,
-    schema,
-    onError: (error) => {
-      console.error('Error submitting request:', error)
-      if (error.message.includes('limit')) {
-        setIsRateLimited(true)
-      }
-
-      setErrorMessage(error.message)
-    },
-    onFinish: async ({ object: fragment, error }) => {
-      if (!error) {
-        // send it to /api/sandbox
-        setIsPreviewLoading(true)
-        posthog.capture('fragment_generated', {
-          template: fragment?.template,
-        })
-
-        const response = await fetch('/api/sandbox', {
-          method: 'POST',
-          body: JSON.stringify({
-            fragment,
-            userID: session?.user?.id,
-            teamID: userTeam?.id,
-            accessToken: session?.access_token,
-          }),
-        })
-
-        const result = await response.json()
-        posthog.capture('sandbox_created', { url: result.url })
-
-        setResult(result)
-        setCurrentPreview({ fragment, result })
-        setMessage({ result })
-        setCurrentTab('fragment')
-        setIsPreviewLoading(false)
-        await saveProject({ fragment, result })
-      }
-    },
-  })
-
-  useEffect(() => {
-    if (object) {
-      setFragment(object)
-      const content: Message['content'] = [
-        { type: 'text', text: object.commentary || '' },
-        { type: 'code', text: object.code || '' },
-      ]
-
-      if (!lastMessage || lastMessage.role !== 'assistant') {
-        addMessage({
-          role: 'assistant',
-          content,
-          object,
-        })
-      }
-
-      if (lastMessage && lastMessage.role === 'assistant') {
-        setMessage({
-          content,
-          object,
-        })
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [object])
-
-  useEffect(() => {
-    if (error) stop()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [error])
-
-  function setMessage(message: Partial<Message>, index?: number) {
-    setMessages((previousMessages) => {
-      const updatedMessages = [...previousMessages]
-      updatedMessages[index ?? previousMessages.length - 1] = {
-        ...previousMessages[index ?? previousMessages.length - 1],
-        ...message,
-      }
-
-      return updatedMessages
-    })
-  }
-
-  async function handleSubmitAuth(e: React.FormEvent<HTMLFormElement>) {
-    e.preventDefault()
-
-    if (!session) {
-      return setAuthDialog(true)
-    }
-
-    if (isLoading) {
-      stop()
-    }
-
-    const content: Message['content'] = []
-    const images = await toMessageImage(files)
-    const pdfs = await toMessageFile(pdfFiles)
-
-    // Only add text content if there's actual text
-    if (chatInput.trim()) {
-      content.push({ type: 'text', text: chatInput })
-    }
-
-    if (images.length > 0) {
-      images.forEach((image) => {
-        content.push({ type: 'image', image })
-      })
-    }
-
-    if (pdfs.length > 0) {
-      pdfs.forEach((pdf) => {
-        content.push(pdf)
-      })
-    }
-
-    const updatedMessages = addMessage({
-      role: 'user',
-      content,
-    })
-
-    submit({
-      userID: session?.user?.id,
-      teamID: userTeam?.id,
-      messages: toAISDKMessages(updatedMessages),
-      template: currentTemplate,
-      model: currentModel,
-      config: languageModel,
-      ...(shouldUseMorph && fragment ? { currentFragment: fragment } : {}),
-    })
-
-    setChatInput('')
-    setFiles([])
-    setPdfFiles([])
-    setCurrentTab('code')
-
-    posthog.capture('chat_submit', {
-      template: selectedTemplate,
-      model: languageModel.model,
-    })
-  }
-
-  function retry() {
-    submit({
-      userID: session?.user?.id,
-      teamID: userTeam?.id,
-      messages: toAISDKMessages(messages),
-      template: currentTemplate,
-      model: currentModel,
-      config: languageModel,
-      ...(shouldUseMorph && fragment ? { currentFragment: fragment } : {}),
-    })
-  }
-
-  function addMessage(message: Message) {
-    setMessages((previousMessages) => [...previousMessages, message])
-    return [...messages, message]
-  }
-
-  function handleSaveInputChange(e: React.ChangeEvent<HTMLTextAreaElement>) {
-    setChatInput(e.target.value)
-  }
-
-  function handleFileChange(change: SetStateAction<File[]>) {
-    setFiles(change)
-  }
-
-  function handlePdfFileChange(change: SetStateAction<File[]>) {
-    setPdfFiles(change)
-  }
+  const { session } = useAuth(setAuthDialog, setAuthView)
 
   function logout() {
     if (supabase) {
@@ -327,249 +18,6 @@ export default function Home() {
     }
   }
 
-  function handleLanguageModelChange(e: LLMModelConfig) {
-    setLanguageModel({ ...languageModel, ...e })
-  }
-
-  function handleSocialClick(target: 'github' | 'x' | 'discord') {
-    if (target === 'github') {
-      window.open('https://github.com/michaltakac/explorable-research', '_blank')
-    } else if (target === 'x') {
-      window.open('https://x.com/michaltakac', '_blank')
-    } else if (target === 'discord') {
-      window.open('https://discord.gg/e2b', '_blank')
-    }
-
-    posthog.capture(`${target}_click`)
-  }
-
-  function handleClearChat() {
-    stop()
-    setChatInput('')
-    setFiles([])
-    setPdfFiles([])
-    setMessages([])
-    setFragment(undefined)
-    setResult(undefined)
-    setCurrentTab('code')
-    setIsPreviewLoading(false)
-  }
-
-  function setCurrentPreview(preview: {
-    fragment: DeepPartial<FragmentSchema> | undefined
-    result: ExecutionResult | undefined
-  }) {
-    setFragment(preview.fragment)
-    setResult(preview.result)
-  }
-
-  function handleUndo() {
-    setMessages((previousMessages) => [...previousMessages.slice(0, -2)])
-    setCurrentPreview({ fragment: undefined, result: undefined })
-  }
-
-  function handleGetStarted() {
-    setShowChatInterface(true)
-    // Scroll to chat section after a small delay to allow state update
-    setTimeout(() => {
-      chatSectionRef.current?.scrollIntoView({ behavior: 'smooth' })
-    }, 100)
-  }
-
-  function handleBackToLanding() {
-    setShowChatInterface(false)
-    handleClearChat()
-    window.scrollTo({ top: 0, behavior: 'smooth' })
-  }
-
-  // Show chat view only when fragment is ready (generation complete)
-  // During generation (isLoading), stay on creator view to show progress there
-  if (fragment || (messages.length > 0 && !isLoading)) {
-    return (
-      <main className="flex min-h-screen max-h-screen">
-        {supabase && (
-          <AuthDialog
-            open={isAuthDialogOpen}
-            setOpen={setAuthDialog}
-            view={authView}
-            supabase={supabase}
-          />
-        )}
-        <div className="grid w-full md:grid-cols-2">
-          <div
-            className={`flex flex-col w-full max-h-full max-w-[800px] mx-auto px-4 overflow-auto ${fragment ? 'col-span-1' : 'col-span-2'}`}
-          >
-            <NavBar
-              session={session}
-              showLogin={() => setAuthDialog(true)}
-              signOut={logout}
-              onSocialClick={handleSocialClick}
-              onClear={handleClearChat}
-              canClear={messages.length > 0}
-              canUndo={messages.length > 1 && !isLoading}
-              onUndo={handleUndo}
-              showGitHubStar={true}
-            />
-            <Chat
-              messages={messages}
-              isLoading={isLoading}
-              setCurrentPreview={setCurrentPreview}
-            />
-            <ChatInput
-              retry={retry}
-              isErrored={error !== undefined}
-              errorMessage={errorMessage}
-              isLoading={isLoading}
-              isRateLimited={isRateLimited}
-              stop={stop}
-              input={chatInput}
-              handleInputChange={handleSaveInputChange}
-              handleSubmit={handleSubmitAuth}
-              isMultiModal={currentModel?.multiModal || false}
-              files={files}
-              handleFileChange={handleFileChange}
-              pdfFiles={pdfFiles}
-              handlePdfFileChange={handlePdfFileChange}
-            >
-              <ChatPicker
-                templates={templates}
-                selectedTemplate={selectedTemplate}
-                onSelectedTemplateChange={setSelectedTemplate}
-                models={filteredModels}
-                languageModel={languageModel}
-                onLanguageModelChange={handleLanguageModelChange}
-              />
-            </ChatInput>
-          </div>
-          <Preview
-            teamID={userTeam?.id}
-            accessToken={session?.access_token}
-            selectedTab={currentTab}
-            onSelectedTabChange={setCurrentTab}
-            isChatLoading={isLoading}
-            isPreviewLoading={isPreviewLoading}
-            fragment={fragment}
-            result={result as ExecutionResult}
-            onClose={() => setFragment(undefined)}
-          />
-        </div>
-      </main>
-    )
-  }
-
-  // Creator view - centered PDF upload interface (after clicking Get Started)
-  if (showChatInterface) {
-    return (
-      <main className="min-h-screen flex flex-col bg-background">
-        {supabase && (
-          <AuthDialog
-            open={isAuthDialogOpen}
-            setOpen={setAuthDialog}
-            view={authView}
-            supabase={supabase}
-          />
-        )}
-        
-        {/* Navbar */}
-        <div className="w-full max-w-4xl mx-auto px-4">
-          <NavBar
-            session={session}
-            showLogin={() => setAuthDialog(true)}
-            signOut={logout}
-            onSocialClick={handleSocialClick}
-            onClear={handleClearChat}
-            canClear={false}
-            canUndo={false}
-            onUndo={handleUndo}
-            showGitHubStar={true}
-          />
-        </div>
-
-        {/* Centered content */}
-        <div className="flex-1 flex items-center justify-center px-4 py-4">
-          <div className="w-full max-w-2xl" ref={chatSectionRef}>
-            {/* Back button - hide when loading */}
-            {!isLoading && (
-              <div className="flex justify-center mb-4">
-                <button
-                  onClick={handleBackToLanding}
-                  className="flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground transition-colors"
-                >
-                  <ArrowLeft className="w-4 h-4" />
-                  Back to home
-                </button>
-              </div>
-            )}
-
-            {/* Headline - hide when loading since CreatorInput has its own */}
-            {!isLoading && (
-              <div className="text-center mb-6">
-                <h1 className="text-2xl sm:text-3xl font-bold text-foreground mb-3">
-                  Create your explorable
-                </h1>
-                <p className="text-muted-foreground">
-                  Upload a research article and we{"'"}ll transform it into an interactive website
-                </p>
-              </div>
-            )}
-
-            {/* Creator input */}
-            <CreatorInput
-              retry={retry}
-              isErrored={error !== undefined}
-              errorMessage={errorMessage}
-              isLoading={isLoading}
-              isRateLimited={isRateLimited}
-              stop={stop}
-              input={chatInput}
-              handleInputChange={handleSaveInputChange}
-              handleSubmit={handleSubmitAuth}
-              isMultiModal={currentModel?.multiModal || false}
-              files={files}
-              handleFileChange={handleFileChange}
-              pdfFiles={pdfFiles}
-              handlePdfFileChange={handlePdfFileChange}
-              selectedModel={languageModel.model || ''}
-            >
-              <ChatPicker
-                templates={templates}
-                selectedTemplate={selectedTemplate}
-                onSelectedTemplateChange={setSelectedTemplate}
-                models={filteredModels}
-                languageModel={languageModel}
-                onLanguageModelChange={handleLanguageModelChange}
-              />
-            </CreatorInput>
-          </div>
-        </div>
-
-        {/* Footer */}
-        <footer className="py-6 border-t border-border/50">
-          <div className="max-w-4xl mx-auto px-4">
-            <p className="text-xs text-muted-foreground text-center">
-              <a href="https://github.com/michaltakac/explorable-research" target="_blank" className="text-violet-600 dark:text-violet-400 hover:underline">
-                Explorable Research
-              </a>
-              {' by '}
-              <a href="https://github.com/michaltakac" target="_blank" className="text-violet-600 dark:text-violet-400 hover:underline">
-                Michal Takáč
-              </a>
-              {' · Based on '}
-              <a href="https://github.com/e2b-dev/fragments" target="_blank" className="text-violet-600 dark:text-violet-400 hover:underline">
-                Fragments
-              </a>
-              {' by '}
-              <a href="https://e2b.dev" target="_blank" className="text-violet-600 dark:text-violet-400 hover:underline">
-                ✶ E2B
-              </a>
-            </p>
-          </div>
-        </footer>
-      </main>
-    )
-  }
-
-  // Landing page view
   return (
     <main className="min-h-screen bg-background">
       {supabase && (
@@ -581,15 +29,15 @@ export default function Home() {
         />
       )}
       
-      <Header onGetStarted={handleGetStarted} session={session} signOut={logout} />
+      <Header session={session} signOut={logout} />
       
-      <Hero onGetStarted={handleGetStarted} />
+      <Hero />
       
       <Features />
       
       <HowItWorks />
       
-      <CTASection onGetStarted={handleGetStarted} />
+      <CTASection />
       
       <Footer />
     </main>
